merge developer to main

+++ src/services/rents/deleteRent.service.ts
@@ -1,8 +1,15 @@
 import { Request } from "express";
-import rentRepository from "../../repositories/rent.repository";
+import { ErrorHandler } from "../../errors/error";
+import { rentRepository } from "../../repositories";
 
 const deleteRentService = async ({ params }: Request) => {
   const { id } = params;
+
+  const rentToDelete = await rentRepository.retrieve({ rentUuid: id });
+
+  if (!rentToDelete) {
+    throw new ErrorHandler(404, "Rent not found");
+  }
 
   await rentRepository.delete(id);
